Migrated Quiz Modal test to TS
